Updated opp_file_get_all_opponent_names to return length of array.

+++ G2ME.h
@@ -82,7 +82,7 @@ struct record *get_all_records(char *, long *);
 
 /* Random helper functions */
 long longest_name(char *, int);
-int filter_player_list(char **, int *, char *);
-int filter_player_list_min_events(char **, int *);
+int filter_player_list(char **, short *, char *);
+int filter_player_list_min_events(char **, short *);
 
 #endif
